reduce duplicated compiler code
